BASIC SALES DASHBOARD
🌟 My First Sales Dashboard Using Power BI! Taking my first steps into data visualization, I created this simple yet insightful sales dashboard using Power BI. It showcases key metrics like sales, target sales trends, and about salesperson distribution all in a minimal design.
This journey has been an exciting opportunity to turn data into meaningful insights, and I’m eager to refine my skills further. Feedback and suggestions are always welcome as I continue to grow in this space. Looking forward to connecting with fellow data enthusiasts .It helps us to understand about:-
~𝘾𝙍𝙀𝘼𝙏𝙄𝙑𝙄𝙏𝙔
~𝙎𝙏𝙊𝙍𝙔𝙏𝙀𝙇𝙇𝙄𝙉𝙂
~𝘼𝙀𝙎𝙏𝙃𝙀𝙏𝙄𝘾𝙎 𝙀𝙏𝘾.